+++ crates/api_client/src/models/initial_profile_age.rs
@@ -14,13 +14,13 @@ use serde::{Deserialize, Serialize};
 #[derive(Clone, Default, Debug, PartialEq, Serialize, Deserialize)]
 pub struct InitialProfileAge {
     #[serde(rename = "initial_profile_age")]
-    pub initial_profile_age: i64,
+    pub initial_profile_age: i32,
     #[serde(rename = "initial_profile_age_set_unix_time")]
     pub initial_profile_age_set_unix_time: Box<models::UnixTime>,
 }
 
 impl InitialProfileAge {
-    pub fn new(initial_profile_age: i64, initial_profile_age_set_unix_time: models::UnixTime) -> InitialProfileAge {
+    pub fn new(initial_profile_age: i32, initial_profile_age_set_unix_time: models::UnixTime) -> InitialProfileAge {
         InitialProfileAge {
             initial_profile_age,
             initial_profile_age_set_unix_time: Box::new(initial_profile_age_set_unix_time),
