+++ crates/api_client/src/models/report_account_info.rs
@@ -14,14 +14,14 @@ use serde::{Deserialize, Serialize};
 #[derive(Clone, Default, Debug, PartialEq, Serialize, Deserialize)]
 pub struct ReportAccountInfo {
     #[serde(rename = "age")]
-    pub age: i64,
+    pub age: i32,
     /// A string wrapper that ensures the string is not empty. This type is used for TEXT columns that should not allow empty strings. In the database, these columns are NULL when there is no value, and this type represents non-NULL values that must be non-empty.
     #[serde(rename = "name", skip_serializing_if = "Option::is_none")]
     pub name: Option<String>,
 }
 
 impl ReportAccountInfo {
-    pub fn new(age: i64) -> ReportAccountInfo {
+    pub fn new(age: i32) -> ReportAccountInfo {
         ReportAccountInfo {
             age,
             name: None,
